- Implement code structure using composition by creating concerns - Add NotFound module to handle ActionController routing error for 404s - Add DeviseWhitelist module for adding allowed attributes for sign-up and account update
