+++ app/controllers/application_controller.rb
@@ -1,15 +1,4 @@
 class ApplicationController < ActionController::Base
-  before_action :configure_permitted_parameters, if: :devise_controller?
-  rescue_from ActiveRecord::RecordNotFound, with: :not_found
-
-  def not_found
-    render file: 'public/404.html', status: :not_found
-  end
-
-  protected
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:name])
-    devise_parameter_sanitizer.permit(:account_update, keys: [:name])
-  end
+  include NotFound
+  include DeviseWhitelist
 end
