+++ app/views/layouts/application.html.erb
@@ -14,7 +14,8 @@
     <p class="notice"><%= notice %></p>
     <p class="alert"><%= alert %></p>
     <% if current_user %>
-      <%= link_to 'Logout', destroy_user_session_path, method: :delete %>
+      <%= link_to 'Logout', destroy_user_session_path, method: :delete %> |
+      <%= link_to 'Update Account', edit_user_registration_path %>
     <% elsif  request.fullpath == root_path %>
       <%= link_to 'Login', new_user_session_path %> |
       <%= link_to 'Register', new_user_registration_path %>
